backend/requirements.txt
fastapi
uvicorn
jsf
python-dotenv
azure-cosmos